requirements.txt
boto3==1.1.1
csvkit==0.9.1
invoke
pandas
# intentionally leaving pandas version blank, b/c `pip install` generally
# fails on Linux - must use the package manager version
requests==2.7.0
gunicorn==19.3.0
cryptography>=1.0
Flask==0.10.1
Flask-SQLAlchemy==2.1
Flask-Cors==2.1.0
Flask-BasicAuth==0.2.0
marshmallow==2.0.0.b4
marshmallow-sqlalchemy==0.4.0
smore==0.2.0
cfenv==0.5.2
pytest
pytest-flask
pytest-cov

git+https://github.com/18f/sandman2@table-delete
git+https://github.com/marshmallow-code/smore@dev

python-dotenv==0.5.1
